# api/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Proxy API
  version: 0.6.0
  description: |
    HTTP-прокси над бесплатными LLM-источниками (Gemini, Groq, OpenRouter).

    ## Для основной системы
    1. `GET /v1/providers` — здоровье источников и квоты пулов  
    2. `GET /v1/models` — модели уже отсортированы: `rank=1` самая сильная  
    3. `POST /v1/route` или `POST /v1/chat/completions` с `model:"auto"` — прокси сам выберет модель  
    4. Либо явный `model` — при 429/5xx прокси сделает fallback (лимит попыток / timeout)  
    5. `GET /v1/stats/summary` — статистика из PostgreSQL  

    ## Auth (опционально)
    Если задан `PROXY_API_KEY`, для `/v1/*` нужен `Authorization: Bearer <key>` или `X-API-Key`.
    Публичны: `/healthz`, `/docs`, `/openapi.yaml`, `/architecture`.  

    ## Квоты
    - `quota.source=live` — live remaining (OpenRouter)  
    - `quota.source=live_headers` — RPM/TPM из upstream rate-limit headers  
    - `quota.source=local_db` — остаток из PostgreSQL (успешные запросы за день)  
    - `quota.source=static_catalog` — только потолок из каталога  
    - `scope=shared_pool` + `pool_id` — общий дневной бюджет на группу моделей  

    ## Ранжирование
    - `quality_score` — эвристика «мозгов» (выше = лучше)  
    - `rank` — позиция после сортировки  

    Архитектура: `docs/architecture.md`  
    Интерактивная документация: [/docs](/docs)
  contact:
    name: llm-proxy
servers:
  - url: http://localhost:8080
    description: Local
tags:
  - name: system
  - name: catalog
  - name: inference
  - name: stats
  - name: docs

paths:
  /healthz:
    get:
      tags: [system]
      summary: Healthcheck
      operationId: healthz
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /docs:
    get:
      tags: [docs]
      summary: Swagger UI
      operationId: docsUI
      responses:
        "200":
          description: HTML Swagger UI
          content:
            text/html:
              schema:
                type: string

  /architecture:
    get:
      tags: [docs]
      summary: Architecture docs with Mermaid diagrams
      description: |
        HTML-страница рендерит `docs/architecture.md` через marked + mermaid.js.
        Встроенный Markdown Preview в Cursor/VS Code Mermaid не рисует — используй этот URL.
      operationId: architectureUI
      responses:
        "200":
          description: HTML architecture viewer
          content:
            text/html:
              schema:
                type: string

  /architecture.md:
    get:
      tags: [docs]
      summary: Raw architecture markdown
      operationId: architectureMarkdown
      responses:
        "200":
          description: Markdown source
          content:
            text/markdown:
              schema:
                type: string

  /openapi.yaml:
    get:
      tags: [docs]
      summary: OpenAPI specification
      operationId: openAPISpec
      responses:
        "200":
          description: OpenAPI 3 YAML
          content:
            application/yaml:
              schema:
                type: string

  /admin/status:
    get:
      tags: [system]
      summary: UI/runtime status
      description: Setup flag, masked keys, gateway defaults.
      operationId: adminStatus
      responses:
        "200":
          description: Status snapshot
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  /admin/keys:
    put:
      tags: [system]
      summary: Save provider API keys (runtime volume) and optionally apply
      operationId: adminKeys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gemini_api_key: { type: string }
                groq_api_key: { type: string }
                openrouter_api_key: { type: string }
                apply: { type: boolean, default: true }
      responses:
        "200":
          description: Updated status
        "400":
          $ref: "#/components/responses/ProxyError"

  /admin/defaults:
    put:
      tags: [system]
      summary: Save default model/sampling for gateway UI
      operationId: adminDefaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Saved defaults

  /admin/apply:
    post:
      tags: [system]
      summary: Hot-reload providers from runtime + env keys
      operationId: adminApply
      responses:
        "200":
          description: Reloaded status

  /v1/providers:
    get:
      tags: [catalog]
      summary: Список провайдеров (источников)
      description: |
        Обзор подключённых backend-ов: health, число моделей, квота пула (если есть).
      operationId: listProviders
      responses:
        "200":
          description: Providers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersResponse"
        "502":
          $ref: "#/components/responses/ProxyError"

  /v1/models:
    get:
      tags: [catalog]
      summary: Список моделей с source, quota и ranking
      description: |
        OpenAI-совместимый список, расширенный полями для роутинга.
        Результат отсортирован по `quality_score` DESC (`rank=1` — лучшая модель в выдаче).
      operationId: listModels
      parameters:
        - name: provider
          in: query
          schema:
            type: string
            enum: [gemini, groq, openrouter]
          description: Фильтр по источнику
        - name: free
          in: query
          schema:
            type: boolean
          description: Только free-модели
        - name: recommended
          in: query
          schema:
            type: boolean
          description: Только рекомендованные для free-роутинга
      responses:
        "200":
          description: Models list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
        "502":
          $ref: "#/components/responses/ProxyError"

  /v1/chat/completions:
    post:
      tags: [inference]
      summary: Chat completions (OpenAI-compatible)
      description: |
        Проксирует запрос в backend по `model`.
        `model` можно не указывать или поставить `"auto"` — прокси выберет лучшую доступную модель
        (rank + remaining) и при 429/5xx пойдёт по fallback.
        Основные sampling-параметры (`temperature`, `top_p`, `max_tokens`, …) принимаются;
        перед upstream тело санитизируется под возможности провайдера (неподдерживаемые поля отбрасываются).
        Алиасы: `POST /v1/completions`, `POST /v1/completion`.
        Ответные заголовки: `X-LLM-Proxy-Model`, `X-LLM-Proxy-Provider`, `X-LLM-Proxy-Attempts`.
      operationId: chatCompletions
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/ProxyAPIKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
            examples:
              no_model:
                summary: Без model — auto-route
                value:
                  messages:
                    - role: user
                      content: ping
                  temperature: 0.7
                  max_tokens: 128
              auto:
                value:
                  model: auto
                  messages:
                    - role: user
                      content: ping
              gemini:
                value:
                  model: gemini-3.5-flash-lite
                  messages:
                    - role: user
                      content: ping
                  temperature: 0.5
                  top_p: 0.9
                  max_tokens: 256
              groq:
                value:
                  model: openai/gpt-oss-20b
                  messages:
                    - role: user
                      content: ping
              openrouter:
                value:
                  model: openrouter/free
                  messages:
                    - role: user
                      content: ping
                  top_k: 40
      responses:
        "200":
          description: Upstream chat completion
          headers:
            X-Request-ID:
              schema: { type: string }
              description: Proxy request id (echo client header or generated)
            X-LLM-Proxy-Model:
              schema: { type: string }
            X-LLM-Proxy-Provider:
              schema: { type: string }
            X-LLM-Proxy-Attempts:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "400":
          $ref: "#/components/responses/ProxyError"
        "404":
          $ref: "#/components/responses/ProxyError"
        "429":
          description: Rate limit (upstream or free tier)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "502":
          $ref: "#/components/responses/ProxyError"
        "503":
          $ref: "#/components/responses/ProxyError"

  /v1/completions:
    post:
      tags: [inference]
      summary: Alias of /v1/chat/completions
      description: То же тело и поведение, что у `/v1/chat/completions` (включая отсутствие `model`).
      operationId: completionsAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Upstream chat completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "400":
          $ref: "#/components/responses/ProxyError"
        "503":
          $ref: "#/components/responses/ProxyError"

  /v1/completion:
    post:
      tags: [inference]
      summary: Alias of /v1/chat/completions
      description: Укороченный алиас для клиентов, которым лень указывать model.
      operationId: completionAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Upstream chat completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "400":
          $ref: "#/components/responses/ProxyError"
        "503":
          $ref: "#/components/responses/ProxyError"

  /v1/route:
    post:
      tags: [inference]
      summary: Auto-route chat to best available model
      description: |
        То же что chat с `model:auto`, плюс фильтры выбора кандидатов.
        Proxy fields (`provider`, `recommended_only`, `exclude_models`) не уходят upstream.
      operationId: routeChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
            example:
              messages:
                - role: user
                  content: ping
              recommended_only: true
      responses:
        "200":
          description: Upstream chat completion from chosen model
          headers:
            X-LLM-Proxy-Model:
              schema: { type: string }
            X-LLM-Proxy-Provider:
              schema: { type: string }
            X-LLM-Proxy-Attempts:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "400":
          $ref: "#/components/responses/ProxyError"
        "503":
          $ref: "#/components/responses/ProxyError"

  /v1/stats/summary:
    get:
      tags: [stats]
      summary: Агрегированная статистика из PostgreSQL
      description: |
        Для поддержки и дашбордов основной системы: успехи, ошибки, 429, токены.
        По умолчанию окно — последние 7 UTC-дней.
      operationId: statsSummary
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
            example: "2026-09-28"
          description: Начало периода (YYYY-MM-DD), default = today-7d UTC
        - name: to
          in: query
          schema:
            type: string
            format: date
            example: "2026-10-04"
          description: Конец периода (YYYY-MM-DD), default = today UTC
        - name: provider
          in: query
          schema:
            type: string
            enum: [gemini, groq, openrouter]
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
        "500":
          $ref: "#/components/responses/ProxyError"

components:
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Опциональный корреляционный id запроса. Если не передан — прокси сгенерирует UUID.
        Возвращается в том же заголовке. Пишется во все логи (`req_id=`) и в `usage_events.request_id`.
      schema:
        type: string
        maxLength: 128
        pattern: '^[A-Za-z0-9._-]{1,128}$'
        example: my-debug-42
    ProxyAPIKey:
      name: Authorization
      in: header
      required: false
      description: |
        Если на сервере задан `PROXY_API_KEY` — обязателен для `/v1/*`.
        Формат: `Bearer <PROXY_API_KEY>`. Альтернатива: заголовок `X-API-Key`.
      schema:
        type: string
        example: Bearer my-secret

  schemas:
    Source:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          example: groq
        name:
          type: string
          example: Groq (GroqCloud)

    Quota:
      type: object
      description: Бюджет запросов/токенов для роутинга
      properties:
        rpm:
          type: integer
          nullable: true
        rpd:
          type: integer
          nullable: true
        tpm:
          type: integer
          nullable: true
        tpd:
          type: integer
          nullable: true
        remaining_rpm:
          type: integer
          nullable: true
        remaining_tpm:
          type: integer
          nullable: true
        remaining_rpd:
          type: integer
          nullable: true
          description: Остаток дневных запросов (live или local_db)
        used_rpd:
          type: integer
          nullable: true
        reset_rpd_hint:
          type: string
          enum: [midnight_utc, midnight_pacific, sliding_window, unknown]
        resets_at:
          type: string
          format: date-time
          description: Следующий сброс RPD (если вычисляем)
        scope:
          type: string
          enum: [per_model, shared_pool]
        pool_id:
          type: string
          example: openrouter:free
        source:
          type: string
          enum: [live, live_headers, local_db, static_catalog, unknown]
          description: live=OpenRouter /key; live_headers=RPM/TPM headers; local_db=PostgreSQL counters; static_catalog=только потолок
        confidence:
          type: string
          enum: [exact, estimate, unknown]
        tier:
          type: string
          example: free
        notes:
          type: string

    Model:
      type: object
      required: [id, object, owned_by, provider, source, free, quota]
      properties:
        id:
          type: string
          example: openai/gpt-oss-20b
        object:
          type: string
          example: model
        created:
          type: integer
          format: int64
        owned_by:
          type: string
        provider:
          type: string
          description: Дублирует source.id для удобства фильтров
          example: groq
        source:
          $ref: "#/components/schemas/Source"
        free:
          type: boolean
        recommended:
          type: boolean
          description: Рекомендована для free-роутинга
        quality_score:
          type: integer
          description: Эвристика качества/«мозгов» (выше = предпочтительнее)
          example: 7400
        rank:
          type: integer
          description: Позиция после сортировки (1 = лучшая в текущей выдаче)
          example: 1
        quota:
          $ref: "#/components/schemas/Quota"

    ModelsResponse:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          example: list
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"

    ProviderStatus:
      type: object
      required: [id, name, enabled, healthy, models, free_models]
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        healthy:
          type: boolean
        error:
          type: string
        models:
          type: integer
        free_models:
          type: integer
        docs:
          type: string
          format: uri
        quota:
          $ref: "#/components/schemas/Quota"

    ProvidersResponse:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          example: list
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProviderStatus"

    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string

    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
          description: |
            ID из GET /v1/models. Если опущен / пустой / `auto` — прокси выбирает лучшую модель
            и делает fallback при 429/5xx.
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        temperature:
          type: number
          minimum: 0
          maximum: 2
        top_p:
          type: number
          minimum: 0
          maximum: 1
        top_k:
          type: integer
          description: Форвардится только если upstream поддерживает (сейчас OpenRouter)
        max_tokens:
          type: integer
          minimum: 1
        max_completion_tokens:
          type: integer
          minimum: 1
          description: Алиас/альтернатива max_tokens (OpenAI-совместимые бэкенды)
        n:
          type: integer
          minimum: 1
          description: На Groq принимается только 1; иначе поле отбрасывается
        stop:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        stream:
          type: boolean
          description: Streaming пока не гарантируется прокси
        presence_penalty:
          type: number
        frequency_penalty:
          type: number
        seed:
          type: integer
        user:
          type: string
        response_format:
          type: object
          additionalProperties: true
        tools:
          type: array
          items:
            type: object
            additionalProperties: true
        tool_choice:
          description: string или object
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
        parallel_tool_calls:
          type: boolean
        reasoning_effort:
          type: string
          description: Gemini / часть OpenRouter; на Groq отбрасывается
        modalities:
          type: array
          items: { type: string }
          description: В основном Gemini OpenAI-compat
        logit_bias:
          type: object
          additionalProperties: true
          description: Отбрасывается на Groq/Gemini
        logprobs:
          type: boolean
          description: Отбрасывается на Groq/Gemini
        top_logprobs:
          type: integer
          description: Отбрасывается на Groq/Gemini

    RouteRequest:
      allOf:
        - $ref: "#/components/schemas/ChatCompletionRequest"
        - type: object
          properties:
            provider:
              type: string
              description: Ограничить кандидатов источником
            recommended_only:
              type: boolean
            exclude_models:
              type: array
              items:
                type: string

    ChatCompletionResponse:
      type: object
      description: Upstream OpenAI-compatible payload (pass-through)
      additionalProperties: true
      properties:
        id:
          type: string
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            additionalProperties: true
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer

    DayStat:
      type: object
      properties:
        day:
          type: string
        provider:
          type: string
        model:
          type: string
        pool_id:
          type: string
        success_count:
          type: integer
        error_count:
          type: integer
        rate_limit_count:
          type: integer
        tokens_in:
          type: integer
        tokens_out:
          type: integer

    StatsSummary:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        total_success:
          type: integer
        total_errors:
          type: integer
        total_rate_limit:
          type: integer
        tokens_in:
          type: integer
        tokens_out:
          type: integer
        by_provider:
          type: array
          items:
            $ref: "#/components/schemas/DayStat"
        by_model:
          type: array
          items:
            $ref: "#/components/schemas/DayStat"

    ProxyErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
              example: proxy_error

    UpstreamError:
      type: object
      description: Тело ошибки upstream (pass-through) или proxy_error
      additionalProperties: true

  responses:
    ProxyError:
      description: Proxy error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProxyErrorBody"